Reverse sun / moon conditions
Now, the button says what it will do (change color mode to light if sun
or dark if moon), not a representation of what it currently is.

I took the original approach from a random blog site, but I found it
confusing and inconsistent with other buttons, which tell you its
action.