Fix crash in ChatGPT example
We were crashing because the server was returning JSON fields we
didn't know about. This either broke because those fields were
new and not accounted for in old code or because we updated the
serialization library at some point and they changed some default
behavior on us.

For now, we make the JSON serializer a lot less strict to avoid
this issue from happening again in the future.